requirements.txt
langchain
langchain-groq
langchain-community
yfinance
python-dotenv
streamlit
numexpr
